LayoutTests/imported/w3c/web-platform-tests/dom/nodes/Document-characterSet-normalization-expected.txt


PASS Name "utf-8" has label "unicode-1-1-utf-8" (characterSet) 
PASS Name "utf-8" has label "unicode-1-1-utf-8" (inputEncoding) 
PASS Name "utf-8" has label "unicode-1-1-utf-8" (charset) 
PASS Name "utf-8" has label "utf-8" (characterSet) 
PASS Name "utf-8" has label "utf-8" (inputEncoding) 
PASS Name "utf-8" has label "utf-8" (charset) 
PASS Name "utf-8" has label "utf8" (characterSet) 
PASS Name "utf-8" has label "utf8" (inputEncoding) 
PASS Name "utf-8" has label "utf8" (charset) 
PASS Name "utf-8" has label "utf-16" (characterSet) 
PASS Name "utf-8" has label "utf-16" (inputEncoding) 
PASS Name "utf-8" has label "utf-16" (charset) 
PASS Name "utf-8" has label "utf-16le" (characterSet) 
PASS Name "utf-8" has label "utf-16le" (inputEncoding) 
PASS Name "utf-8" has label "utf-16le" (charset) 
PASS Name "utf-8" has label "utf-16be" (characterSet) 
PASS Name "utf-8" has label "utf-16be" (inputEncoding) 
PASS Name "utf-8" has label "utf-16be" (charset) 
PASS Name "ibm866" has label "866" (characterSet) 
PASS Name "ibm866" has label "866" (inputEncoding) 
PASS Name "ibm866" has label "866" (charset) 
PASS Name "ibm866" has label "cp866" (characterSet) 
PASS Name "ibm866" has label "cp866" (inputEncoding) 
PASS Name "ibm866" has label "cp866" (charset) 
PASS Name "ibm866" has label "csibm866" (characterSet) 
PASS Name "ibm866" has label "csibm866" (inputEncoding) 
PASS Name "ibm866" has label "csibm866" (charset) 
PASS Name "ibm866" has label "ibm866" (characterSet) 
PASS Name "ibm866" has label "ibm866" (inputEncoding) 
PASS Name "ibm866" has label "ibm866" (charset) 
PASS Name "iso-8859-2" has label "csisolatin2" (characterSet) 
PASS Name "iso-8859-2" has label "csisolatin2" (inputEncoding) 
PASS Name "iso-8859-2" has label "csisolatin2" (charset) 
PASS Name "iso-8859-2" has label "iso-8859-2" (characterSet) 
PASS Name "iso-8859-2" has label "iso-8859-2" (inputEncoding) 
PASS Name "iso-8859-2" has label "iso-8859-2" (charset) 
PASS Name "iso-8859-2" has label "iso-ir-101" (characterSet) 
PASS Name "iso-8859-2" has label "iso-ir-101" (inputEncoding) 
PASS Name "iso-8859-2" has label "iso-ir-101" (charset) 
PASS Name "iso-8859-2" has label "iso8859-2" (characterSet) 
PASS Name "iso-8859-2" has label "iso8859-2" (inputEncoding) 
PASS Name "iso-8859-2" has label "iso8859-2" (charset) 
FAIL Name "iso-8859-2" has label "iso88592" (characterSet) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso88592" (inputEncoding) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso88592" (charset) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso_8859-2" (characterSet) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso_8859-2" (inputEncoding) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso_8859-2" (charset) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
PASS Name "iso-8859-2" has label "iso_8859-2:1987" (characterSet) 
PASS Name "iso-8859-2" has label "iso_8859-2:1987" (inputEncoding) 
PASS Name "iso-8859-2" has label "iso_8859-2:1987" (charset) 
PASS Name "iso-8859-2" has label "l2" (characterSet) 
PASS Name "iso-8859-2" has label "l2" (inputEncoding) 
PASS Name "iso-8859-2" has label "l2" (charset) 
PASS Name "iso-8859-2" has label "latin2" (characterSet) 
PASS Name "iso-8859-2" has label "latin2" (inputEncoding) 
PASS Name "iso-8859-2" has label "latin2" (charset) 
PASS Name "iso-8859-3" has label "csisolatin3" (characterSet) 
PASS Name "iso-8859-3" has label "csisolatin3" (inputEncoding) 
PASS Name "iso-8859-3" has label "csisolatin3" (charset) 
PASS Name "iso-8859-3" has label "iso-8859-3" (characterSet) 
PASS Name "iso-8859-3" has label "iso-8859-3" (inputEncoding) 
PASS Name "iso-8859-3" has label "iso-8859-3" (charset) 
PASS Name "iso-8859-3" has label "iso-ir-109" (characterSet) 
PASS Name "iso-8859-3" has label "iso-ir-109" (inputEncoding) 
PASS Name "iso-8859-3" has label "iso-ir-109" (charset) 
PASS Name "iso-8859-3" has label "iso8859-3" (characterSet) 
PASS Name "iso-8859-3" has label "iso8859-3" (inputEncoding) 
PASS Name "iso-8859-3" has label "iso8859-3" (charset) 
FAIL Name "iso-8859-3" has label "iso88593" (characterSet) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso88593" (inputEncoding) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso88593" (charset) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso_8859-3" (characterSet) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso_8859-3" (inputEncoding) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso_8859-3" (charset) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
PASS Name "iso-8859-3" has label "iso_8859-3:1988" (characterSet) 
PASS Name "iso-8859-3" has label "iso_8859-3:1988" (inputEncoding) 
PASS Name "iso-8859-3" has label "iso_8859-3:1988" (charset) 
PASS Name "iso-8859-3" has label "l3" (characterSet) 
PASS Name "iso-8859-3" has label "l3" (inputEncoding) 
PASS Name "iso-8859-3" has label "l3" (charset) 
PASS Name "iso-8859-3" has label "latin3" (characterSet) 
PASS Name "iso-8859-3" has label "latin3" (inputEncoding) 
PASS Name "iso-8859-3" has label "latin3" (charset) 
PASS Name "iso-8859-4" has label "csisolatin4" (characterSet) 
PASS Name "iso-8859-4" has label "csisolatin4" (inputEncoding) 
PASS Name "iso-8859-4" has label "csisolatin4" (charset) 
PASS Name "iso-8859-4" has label "iso-8859-4" (characterSet) 
PASS Name "iso-8859-4" has label "iso-8859-4" (inputEncoding) 
PASS Name "iso-8859-4" has label "iso-8859-4" (charset) 
PASS Name "iso-8859-4" has label "iso-ir-110" (characterSet) 
PASS Name "iso-8859-4" has label "iso-ir-110" (inputEncoding) 
PASS Name "iso-8859-4" has label "iso-ir-110" (charset) 
PASS Name "iso-8859-4" has label "iso8859-4" (characterSet) 
PASS Name "iso-8859-4" has label "iso8859-4" (inputEncoding) 
PASS Name "iso-8859-4" has label "iso8859-4" (charset) 
FAIL Name "iso-8859-4" has label "iso88594" (characterSet) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso88594" (inputEncoding) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso88594" (charset) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso_8859-4" (characterSet) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso_8859-4" (inputEncoding) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso_8859-4" (charset) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
PASS Name "iso-8859-4" has label "iso_8859-4:1988" (characterSet) 
PASS Name "iso-8859-4" has label "iso_8859-4:1988" (inputEncoding) 
PASS Name "iso-8859-4" has label "iso_8859-4:1988" (charset) 
PASS Name "iso-8859-4" has label "l4" (characterSet) 
PASS Name "iso-8859-4" has label "l4" (inputEncoding) 
PASS Name "iso-8859-4" has label "l4" (charset) 
PASS Name "iso-8859-4" has label "latin4" (characterSet) 
PASS Name "iso-8859-4" has label "latin4" (inputEncoding) 
PASS Name "iso-8859-4" has label "latin4" (charset) 
PASS Name "iso-8859-5" has label "csisolatincyrillic" (characterSet) 
PASS Name "iso-8859-5" has label "csisolatincyrillic" (inputEncoding) 
PASS Name "iso-8859-5" has label "csisolatincyrillic" (charset) 
PASS Name "iso-8859-5" has label "cyrillic" (characterSet) 
PASS Name "iso-8859-5" has label "cyrillic" (inputEncoding) 
PASS Name "iso-8859-5" has label "cyrillic" (charset) 
PASS Name "iso-8859-5" has label "iso-8859-5" (characterSet) 
PASS Name "iso-8859-5" has label "iso-8859-5" (inputEncoding) 
PASS Name "iso-8859-5" has label "iso-8859-5" (charset) 
PASS Name "iso-8859-5" has label "iso-ir-144" (characterSet) 
PASS Name "iso-8859-5" has label "iso-ir-144" (inputEncoding) 
PASS Name "iso-8859-5" has label "iso-ir-144" (charset) 
PASS Name "iso-8859-5" has label "iso8859-5" (characterSet) 
PASS Name "iso-8859-5" has label "iso8859-5" (inputEncoding) 
PASS Name "iso-8859-5" has label "iso8859-5" (charset) 
FAIL Name "iso-8859-5" has label "iso88595" (characterSet) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso88595" (inputEncoding) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso88595" (charset) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso_8859-5" (characterSet) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso_8859-5" (inputEncoding) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso_8859-5" (charset) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
PASS Name "iso-8859-5" has label "iso_8859-5:1988" (characterSet) 
PASS Name "iso-8859-5" has label "iso_8859-5:1988" (inputEncoding) 
PASS Name "iso-8859-5" has label "iso_8859-5:1988" (charset) 
PASS Name "iso-8859-6" has label "arabic" (characterSet) 
PASS Name "iso-8859-6" has label "arabic" (inputEncoding) 
PASS Name "iso-8859-6" has label "arabic" (charset) 
PASS Name "iso-8859-6" has label "asmo-708" (characterSet) 
PASS Name "iso-8859-6" has label "asmo-708" (inputEncoding) 
PASS Name "iso-8859-6" has label "asmo-708" (charset) 
FAIL Name "iso-8859-6" has label "csiso88596e" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "csiso88596e" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "csiso88596e" (charset) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "csiso88596i" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "csiso88596i" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "csiso88596i" (charset) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
PASS Name "iso-8859-6" has label "csisolatinarabic" (characterSet) 
PASS Name "iso-8859-6" has label "csisolatinarabic" (inputEncoding) 
PASS Name "iso-8859-6" has label "csisolatinarabic" (charset) 
PASS Name "iso-8859-6" has label "ecma-114" (characterSet) 
PASS Name "iso-8859-6" has label "ecma-114" (inputEncoding) 
PASS Name "iso-8859-6" has label "ecma-114" (charset) 
PASS Name "iso-8859-6" has label "iso-8859-6" (characterSet) 
PASS Name "iso-8859-6" has label "iso-8859-6" (inputEncoding) 
PASS Name "iso-8859-6" has label "iso-8859-6" (charset) 
PASS Name "iso-8859-6" has label "iso-8859-6-e" (characterSet) 
PASS Name "iso-8859-6" has label "iso-8859-6-e" (inputEncoding) 
PASS Name "iso-8859-6" has label "iso-8859-6-e" (charset) 
PASS Name "iso-8859-6" has label "iso-8859-6-i" (characterSet) 
PASS Name "iso-8859-6" has label "iso-8859-6-i" (inputEncoding) 
PASS Name "iso-8859-6" has label "iso-8859-6-i" (charset) 
PASS Name "iso-8859-6" has label "iso-ir-127" (characterSet) 
PASS Name "iso-8859-6" has label "iso-ir-127" (inputEncoding) 
PASS Name "iso-8859-6" has label "iso-ir-127" (charset) 
PASS Name "iso-8859-6" has label "iso8859-6" (characterSet) 
PASS Name "iso-8859-6" has label "iso8859-6" (inputEncoding) 
PASS Name "iso-8859-6" has label "iso8859-6" (charset) 
FAIL Name "iso-8859-6" has label "iso88596" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso88596" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso88596" (charset) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso_8859-6" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso_8859-6" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso_8859-6" (charset) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
PASS Name "iso-8859-6" has label "iso_8859-6:1987" (characterSet) 
PASS Name "iso-8859-6" has label "iso_8859-6:1987" (inputEncoding) 
PASS Name "iso-8859-6" has label "iso_8859-6:1987" (charset) 
PASS Name "iso-8859-7" has label "csisolatingreek" (characterSet) 
PASS Name "iso-8859-7" has label "csisolatingreek" (inputEncoding) 
PASS Name "iso-8859-7" has label "csisolatingreek" (charset) 
PASS Name "iso-8859-7" has label "ecma-118" (characterSet) 
PASS Name "iso-8859-7" has label "ecma-118" (inputEncoding) 
PASS Name "iso-8859-7" has label "ecma-118" (charset) 
PASS Name "iso-8859-7" has label "elot_928" (characterSet) 
PASS Name "iso-8859-7" has label "elot_928" (inputEncoding) 
PASS Name "iso-8859-7" has label "elot_928" (charset) 
PASS Name "iso-8859-7" has label "greek" (characterSet) 
PASS Name "iso-8859-7" has label "greek" (inputEncoding) 
PASS Name "iso-8859-7" has label "greek" (charset) 
PASS Name "iso-8859-7" has label "greek8" (characterSet) 
PASS Name "iso-8859-7" has label "greek8" (inputEncoding) 
PASS Name "iso-8859-7" has label "greek8" (charset) 
PASS Name "iso-8859-7" has label "iso-8859-7" (characterSet) 
PASS Name "iso-8859-7" has label "iso-8859-7" (inputEncoding) 
PASS Name "iso-8859-7" has label "iso-8859-7" (charset) 
PASS Name "iso-8859-7" has label "iso-ir-126" (characterSet) 
PASS Name "iso-8859-7" has label "iso-ir-126" (inputEncoding) 
PASS Name "iso-8859-7" has label "iso-ir-126" (charset) 
PASS Name "iso-8859-7" has label "iso8859-7" (characterSet) 
PASS Name "iso-8859-7" has label "iso8859-7" (inputEncoding) 
PASS Name "iso-8859-7" has label "iso8859-7" (charset) 
FAIL Name "iso-8859-7" has label "iso88597" (characterSet) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso88597" (inputEncoding) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso88597" (charset) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso_8859-7" (characterSet) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso_8859-7" (inputEncoding) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso_8859-7" (charset) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
PASS Name "iso-8859-7" has label "iso_8859-7:1987" (characterSet) 
PASS Name "iso-8859-7" has label "iso_8859-7:1987" (inputEncoding) 
PASS Name "iso-8859-7" has label "iso_8859-7:1987" (charset) 
PASS Name "iso-8859-7" has label "sun_eu_greek" (characterSet) 
PASS Name "iso-8859-7" has label "sun_eu_greek" (inputEncoding) 
PASS Name "iso-8859-7" has label "sun_eu_greek" (charset) 
FAIL Name "iso-8859-8" has label "csiso88598e" (characterSet) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "csiso88598e" (inputEncoding) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "csiso88598e" (charset) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
PASS Name "iso-8859-8" has label "csisolatinhebrew" (characterSet) 
PASS Name "iso-8859-8" has label "csisolatinhebrew" (inputEncoding) 
PASS Name "iso-8859-8" has label "csisolatinhebrew" (charset) 
PASS Name "iso-8859-8" has label "hebrew" (characterSet) 
PASS Name "iso-8859-8" has label "hebrew" (inputEncoding) 
PASS Name "iso-8859-8" has label "hebrew" (charset) 
PASS Name "iso-8859-8" has label "iso-8859-8" (characterSet) 
PASS Name "iso-8859-8" has label "iso-8859-8" (inputEncoding) 
PASS Name "iso-8859-8" has label "iso-8859-8" (charset) 
PASS Name "iso-8859-8" has label "iso-8859-8-e" (characterSet) 
PASS Name "iso-8859-8" has label "iso-8859-8-e" (inputEncoding) 
PASS Name "iso-8859-8" has label "iso-8859-8-e" (charset) 
PASS Name "iso-8859-8" has label "iso-ir-138" (characterSet) 
PASS Name "iso-8859-8" has label "iso-ir-138" (inputEncoding) 
PASS Name "iso-8859-8" has label "iso-ir-138" (charset) 
PASS Name "iso-8859-8" has label "iso8859-8" (characterSet) 
PASS Name "iso-8859-8" has label "iso8859-8" (inputEncoding) 
PASS Name "iso-8859-8" has label "iso8859-8" (charset) 
FAIL Name "iso-8859-8" has label "iso88598" (characterSet) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso88598" (inputEncoding) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso88598" (charset) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso_8859-8" (characterSet) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso_8859-8" (inputEncoding) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso_8859-8" (charset) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
PASS Name "iso-8859-8" has label "iso_8859-8:1988" (characterSet) 
PASS Name "iso-8859-8" has label "iso_8859-8:1988" (inputEncoding) 
PASS Name "iso-8859-8" has label "iso_8859-8:1988" (charset) 
PASS Name "iso-8859-8" has label "visual" (characterSet) 
PASS Name "iso-8859-8" has label "visual" (inputEncoding) 
PASS Name "iso-8859-8" has label "visual" (charset) 
PASS Name "iso-8859-8-i" has label "csiso88598i" (characterSet) 
PASS Name "iso-8859-8-i" has label "csiso88598i" (inputEncoding) 
PASS Name "iso-8859-8-i" has label "csiso88598i" (charset) 
PASS Name "iso-8859-8-i" has label "iso-8859-8-i" (characterSet) 
PASS Name "iso-8859-8-i" has label "iso-8859-8-i" (inputEncoding) 
PASS Name "iso-8859-8-i" has label "iso-8859-8-i" (charset) 
PASS Name "iso-8859-8-i" has label "logical" (characterSet) 
PASS Name "iso-8859-8-i" has label "logical" (inputEncoding) 
PASS Name "iso-8859-8-i" has label "logical" (charset) 
PASS Name "iso-8859-10" has label "csisolatin6" (characterSet) 
PASS Name "iso-8859-10" has label "csisolatin6" (inputEncoding) 
PASS Name "iso-8859-10" has label "csisolatin6" (charset) 
PASS Name "iso-8859-10" has label "iso-8859-10" (characterSet) 
PASS Name "iso-8859-10" has label "iso-8859-10" (inputEncoding) 
PASS Name "iso-8859-10" has label "iso-8859-10" (charset) 
PASS Name "iso-8859-10" has label "iso-ir-157" (characterSet) 
PASS Name "iso-8859-10" has label "iso-ir-157" (inputEncoding) 
PASS Name "iso-8859-10" has label "iso-ir-157" (charset) 
PASS Name "iso-8859-10" has label "iso8859-10" (characterSet) 
PASS Name "iso-8859-10" has label "iso8859-10" (inputEncoding) 
PASS Name "iso-8859-10" has label "iso8859-10" (charset) 
FAIL Name "iso-8859-10" has label "iso885910" (characterSet) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "iso885910" (inputEncoding) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "iso885910" (charset) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
PASS Name "iso-8859-10" has label "l6" (characterSet) 
PASS Name "iso-8859-10" has label "l6" (inputEncoding) 
PASS Name "iso-8859-10" has label "l6" (charset) 
PASS Name "iso-8859-10" has label "latin6" (characterSet) 
PASS Name "iso-8859-10" has label "latin6" (inputEncoding) 
PASS Name "iso-8859-10" has label "latin6" (charset) 
PASS Name "iso-8859-13" has label "iso-8859-13" (characterSet) 
PASS Name "iso-8859-13" has label "iso-8859-13" (inputEncoding) 
PASS Name "iso-8859-13" has label "iso-8859-13" (charset) 
PASS Name "iso-8859-13" has label "iso8859-13" (characterSet) 
PASS Name "iso-8859-13" has label "iso8859-13" (inputEncoding) 
PASS Name "iso-8859-13" has label "iso8859-13" (charset) 
FAIL Name "iso-8859-13" has label "iso885913" (characterSet) assert_equals: expected "ISO-8859-13" but got "ISO-8859-1"
FAIL Name "iso-8859-13" has label "iso885913" (inputEncoding) assert_equals: expected "ISO-8859-13" but got "ISO-8859-1"
FAIL Name "iso-8859-13" has label "iso885913" (charset) assert_equals: expected "ISO-8859-13" but got "ISO-8859-1"
PASS Name "iso-8859-14" has label "iso-8859-14" (characterSet) 
PASS Name "iso-8859-14" has label "iso-8859-14" (inputEncoding) 
PASS Name "iso-8859-14" has label "iso-8859-14" (charset) 
PASS Name "iso-8859-14" has label "iso8859-14" (characterSet) 
PASS Name "iso-8859-14" has label "iso8859-14" (inputEncoding) 
PASS Name "iso-8859-14" has label "iso8859-14" (charset) 
FAIL Name "iso-8859-14" has label "iso885914" (characterSet) assert_equals: expected "ISO-8859-14" but got "ISO-8859-1"
FAIL Name "iso-8859-14" has label "iso885914" (inputEncoding) assert_equals: expected "ISO-8859-14" but got "ISO-8859-1"
FAIL Name "iso-8859-14" has label "iso885914" (charset) assert_equals: expected "ISO-8859-14" but got "ISO-8859-1"
PASS Name "iso-8859-15" has label "csisolatin9" (characterSet) 
PASS Name "iso-8859-15" has label "csisolatin9" (inputEncoding) 
PASS Name "iso-8859-15" has label "csisolatin9" (charset) 
PASS Name "iso-8859-15" has label "iso-8859-15" (characterSet) 
PASS Name "iso-8859-15" has label "iso-8859-15" (inputEncoding) 
PASS Name "iso-8859-15" has label "iso-8859-15" (charset) 
PASS Name "iso-8859-15" has label "iso8859-15" (characterSet) 
PASS Name "iso-8859-15" has label "iso8859-15" (inputEncoding) 
PASS Name "iso-8859-15" has label "iso8859-15" (charset) 
FAIL Name "iso-8859-15" has label "iso885915" (characterSet) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "iso885915" (inputEncoding) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "iso885915" (charset) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "iso_8859-15" (characterSet) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "iso_8859-15" (inputEncoding) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "iso_8859-15" (charset) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
PASS Name "iso-8859-15" has label "l9" (characterSet) 
PASS Name "iso-8859-15" has label "l9" (inputEncoding) 
PASS Name "iso-8859-15" has label "l9" (charset) 
PASS Name "iso-8859-16" has label "iso-8859-16" (characterSet) 
PASS Name "iso-8859-16" has label "iso-8859-16" (inputEncoding) 
PASS Name "iso-8859-16" has label "iso-8859-16" (charset) 
PASS Name "koi8-r" has label "cskoi8r" (characterSet) 
PASS Name "koi8-r" has label "cskoi8r" (inputEncoding) 
PASS Name "koi8-r" has label "cskoi8r" (charset) 
PASS Name "koi8-r" has label "koi" (characterSet) 
PASS Name "koi8-r" has label "koi" (inputEncoding) 
PASS Name "koi8-r" has label "koi" (charset) 
PASS Name "koi8-r" has label "koi8" (characterSet) 
PASS Name "koi8-r" has label "koi8" (inputEncoding) 
PASS Name "koi8-r" has label "koi8" (charset) 
PASS Name "koi8-r" has label "koi8-r" (characterSet) 
PASS Name "koi8-r" has label "koi8-r" (inputEncoding) 
PASS Name "koi8-r" has label "koi8-r" (charset) 
FAIL Name "koi8-r" has label "koi8_r" (characterSet) assert_equals: expected "KOI8-R" but got "ISO-8859-1"
FAIL Name "koi8-r" has label "koi8_r" (inputEncoding) assert_equals: expected "KOI8-R" but got "ISO-8859-1"
FAIL Name "koi8-r" has label "koi8_r" (charset) assert_equals: expected "KOI8-R" but got "ISO-8859-1"
PASS Name "koi8-u" has label "koi8-u" (characterSet) 
PASS Name "koi8-u" has label "koi8-u" (inputEncoding) 
PASS Name "koi8-u" has label "koi8-u" (charset) 
FAIL Name "macintosh" has label "csmacintosh" (characterSet) assert_equals: expected "MACINTOSH" but got "macintosh"
FAIL Name "macintosh" has label "csmacintosh" (inputEncoding) assert_equals: expected "MACINTOSH" but got "macintosh"
FAIL Name "macintosh" has label "csmacintosh" (charset) assert_equals: expected "MACINTOSH" but got "macintosh"
FAIL Name "macintosh" has label "mac" (characterSet) assert_equals: expected "MACINTOSH" but got "macintosh"
FAIL Name "macintosh" has label "mac" (inputEncoding) assert_equals: expected "MACINTOSH" but got "macintosh"
FAIL Name "macintosh" has label "mac" (charset) assert_equals: expected "MACINTOSH" but got "macintosh"
FAIL Name "macintosh" has label "macintosh" (characterSet) assert_equals: expected "MACINTOSH" but got "macintosh"
FAIL Name "macintosh" has label "macintosh" (inputEncoding) assert_equals: expected "MACINTOSH" but got "macintosh"
FAIL Name "macintosh" has label "macintosh" (charset) assert_equals: expected "MACINTOSH" but got "macintosh"
FAIL Name "macintosh" has label "x-mac-roman" (characterSet) assert_equals: expected "MACINTOSH" but got "macintosh"
FAIL Name "macintosh" has label "x-mac-roman" (inputEncoding) assert_equals: expected "MACINTOSH" but got "macintosh"
FAIL Name "macintosh" has label "x-mac-roman" (charset) assert_equals: expected "MACINTOSH" but got "macintosh"
FAIL Name "windows-874" has label "dos-874" (characterSet) assert_equals: expected "WINDOWS-874" but got "windows-874"
FAIL Name "windows-874" has label "dos-874" (inputEncoding) assert_equals: expected "WINDOWS-874" but got "windows-874"
FAIL Name "windows-874" has label "dos-874" (charset) assert_equals: expected "WINDOWS-874" but got "windows-874"
FAIL Name "windows-874" has label "iso-8859-11" (characterSet) assert_equals: expected "WINDOWS-874" but got "windows-874"
FAIL Name "windows-874" has label "iso-8859-11" (inputEncoding) assert_equals: expected "WINDOWS-874" but got "windows-874"
FAIL Name "windows-874" has label "iso-8859-11" (charset) assert_equals: expected "WINDOWS-874" but got "windows-874"
FAIL Name "windows-874" has label "iso8859-11" (characterSet) assert_equals: expected "WINDOWS-874" but got "windows-874"
FAIL Name "windows-874" has label "iso8859-11" (inputEncoding) assert_equals: expected "WINDOWS-874" but got "windows-874"
FAIL Name "windows-874" has label "iso8859-11" (charset) assert_equals: expected "WINDOWS-874" but got "windows-874"
FAIL Name "windows-874" has label "iso885911" (characterSet) assert_equals: expected "WINDOWS-874" but got "ISO-8859-1"
FAIL Name "windows-874" has label "iso885911" (inputEncoding) assert_equals: expected "WINDOWS-874" but got "ISO-8859-1"
FAIL Name "windows-874" has label "iso885911" (charset) assert_equals: expected "WINDOWS-874" but got "ISO-8859-1"
FAIL Name "windows-874" has label "tis-620" (characterSet) assert_equals: expected "WINDOWS-874" but got "windows-874"
FAIL Name "windows-874" has label "tis-620" (inputEncoding) assert_equals: expected "WINDOWS-874" but got "windows-874"
FAIL Name "windows-874" has label "tis-620" (charset) assert_equals: expected "WINDOWS-874" but got "windows-874"
FAIL Name "windows-874" has label "windows-874" (characterSet) assert_equals: expected "WINDOWS-874" but got "windows-874"
FAIL Name "windows-874" has label "windows-874" (inputEncoding) assert_equals: expected "WINDOWS-874" but got "windows-874"
FAIL Name "windows-874" has label "windows-874" (charset) assert_equals: expected "WINDOWS-874" but got "windows-874"
FAIL Name "windows-1250" has label "cp1250" (characterSet) assert_equals: expected "WINDOWS-1250" but got "windows-1250"
FAIL Name "windows-1250" has label "cp1250" (inputEncoding) assert_equals: expected "WINDOWS-1250" but got "windows-1250"
FAIL Name "windows-1250" has label "cp1250" (charset) assert_equals: expected "WINDOWS-1250" but got "windows-1250"
FAIL Name "windows-1250" has label "windows-1250" (characterSet) assert_equals: expected "WINDOWS-1250" but got "windows-1250"
FAIL Name "windows-1250" has label "windows-1250" (inputEncoding) assert_equals: expected "WINDOWS-1250" but got "windows-1250"
FAIL Name "windows-1250" has label "windows-1250" (charset) assert_equals: expected "WINDOWS-1250" but got "windows-1250"
FAIL Name "windows-1250" has label "x-cp1250" (characterSet) assert_equals: expected "WINDOWS-1250" but got "windows-1250"
FAIL Name "windows-1250" has label "x-cp1250" (inputEncoding) assert_equals: expected "WINDOWS-1250" but got "windows-1250"
FAIL Name "windows-1250" has label "x-cp1250" (charset) assert_equals: expected "WINDOWS-1250" but got "windows-1250"
FAIL Name "windows-1251" has label "cp1251" (characterSet) assert_equals: expected "WINDOWS-1251" but got "windows-1251"
FAIL Name "windows-1251" has label "cp1251" (inputEncoding) assert_equals: expected "WINDOWS-1251" but got "windows-1251"
FAIL Name "windows-1251" has label "cp1251" (charset) assert_equals: expected "WINDOWS-1251" but got "windows-1251"
FAIL Name "windows-1251" has label "windows-1251" (characterSet) assert_equals: expected "WINDOWS-1251" but got "windows-1251"
FAIL Name "windows-1251" has label "windows-1251" (inputEncoding) assert_equals: expected "WINDOWS-1251" but got "windows-1251"
FAIL Name "windows-1251" has label "windows-1251" (charset) assert_equals: expected "WINDOWS-1251" but got "windows-1251"
FAIL Name "windows-1251" has label "x-cp1251" (characterSet) assert_equals: expected "WINDOWS-1251" but got "windows-1251"
FAIL Name "windows-1251" has label "x-cp1251" (inputEncoding) assert_equals: expected "WINDOWS-1251" but got "windows-1251"
FAIL Name "windows-1251" has label "x-cp1251" (charset) assert_equals: expected "WINDOWS-1251" but got "windows-1251"
FAIL Name "windows-1252" has label "ansi_x3.4-1968" (characterSet) assert_equals: expected "WINDOWS-1252" but got "US-ASCII"
FAIL Name "windows-1252" has label "ansi_x3.4-1968" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "US-ASCII"
FAIL Name "windows-1252" has label "ansi_x3.4-1968" (charset) assert_equals: expected "WINDOWS-1252" but got "US-ASCII"
FAIL Name "windows-1252" has label "ascii" (characterSet) assert_equals: expected "WINDOWS-1252" but got "US-ASCII"
FAIL Name "windows-1252" has label "ascii" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "US-ASCII"
FAIL Name "windows-1252" has label "ascii" (charset) assert_equals: expected "WINDOWS-1252" but got "US-ASCII"
FAIL Name "windows-1252" has label "cp1252" (characterSet) assert_equals: expected "WINDOWS-1252" but got "windows-1252"
FAIL Name "windows-1252" has label "cp1252" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "windows-1252"
FAIL Name "windows-1252" has label "cp1252" (charset) assert_equals: expected "WINDOWS-1252" but got "windows-1252"
FAIL Name "windows-1252" has label "cp819" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "cp819" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "cp819" (charset) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "csisolatin1" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "csisolatin1" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "csisolatin1" (charset) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "ibm819" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "ibm819" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "ibm819" (charset) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso-8859-1" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso-8859-1" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso-8859-1" (charset) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso-ir-100" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso-ir-100" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso-ir-100" (charset) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso8859-1" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso8859-1" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso8859-1" (charset) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso88591" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso88591" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso88591" (charset) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso_8859-1" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso_8859-1" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso_8859-1" (charset) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso_8859-1:1987" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso_8859-1:1987" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso_8859-1:1987" (charset) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "l1" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "l1" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "l1" (charset) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "latin1" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "latin1" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "latin1" (charset) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "us-ascii" (characterSet) assert_equals: expected "WINDOWS-1252" but got "US-ASCII"
FAIL Name "windows-1252" has label "us-ascii" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "US-ASCII"
FAIL Name "windows-1252" has label "us-ascii" (charset) assert_equals: expected "WINDOWS-1252" but got "US-ASCII"
FAIL Name "windows-1252" has label "windows-1252" (characterSet) assert_equals: expected "WINDOWS-1252" but got "windows-1252"
FAIL Name "windows-1252" has label "windows-1252" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "windows-1252"
FAIL Name "windows-1252" has label "windows-1252" (charset) assert_equals: expected "WINDOWS-1252" but got "windows-1252"
FAIL Name "windows-1252" has label "x-cp1252" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "x-cp1252" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "x-cp1252" (charset) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "x-user-defined" (characterSet) assert_equals: expected "WINDOWS-1252" but got "windows-1252"
FAIL Name "windows-1252" has label "x-user-defined" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "windows-1252"
FAIL Name "windows-1252" has label "x-user-defined" (charset) assert_equals: expected "WINDOWS-1252" but got "windows-1252"
FAIL Name "windows-1253" has label "cp1253" (characterSet) assert_equals: expected "WINDOWS-1253" but got "windows-1253"
FAIL Name "windows-1253" has label "cp1253" (inputEncoding) assert_equals: expected "WINDOWS-1253" but got "windows-1253"
FAIL Name "windows-1253" has label "cp1253" (charset) assert_equals: expected "WINDOWS-1253" but got "windows-1253"
FAIL Name "windows-1253" has label "windows-1253" (characterSet) assert_equals: expected "WINDOWS-1253" but got "windows-1253"
FAIL Name "windows-1253" has label "windows-1253" (inputEncoding) assert_equals: expected "WINDOWS-1253" but got "windows-1253"
FAIL Name "windows-1253" has label "windows-1253" (charset) assert_equals: expected "WINDOWS-1253" but got "windows-1253"
FAIL Name "windows-1253" has label "x-cp1253" (characterSet) assert_equals: expected "WINDOWS-1253" but got "ISO-8859-1"
FAIL Name "windows-1253" has label "x-cp1253" (inputEncoding) assert_equals: expected "WINDOWS-1253" but got "ISO-8859-1"
FAIL Name "windows-1253" has label "x-cp1253" (charset) assert_equals: expected "WINDOWS-1253" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "cp1254" (characterSet) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "cp1254" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "cp1254" (charset) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "csisolatin5" (characterSet) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "csisolatin5" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "csisolatin5" (charset) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "iso-8859-9" (characterSet) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "iso-8859-9" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "iso-8859-9" (charset) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "iso-ir-148" (characterSet) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "iso-ir-148" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "iso-ir-148" (charset) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "iso8859-9" (characterSet) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "iso8859-9" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "iso8859-9" (charset) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "iso88599" (characterSet) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso88599" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso88599" (charset) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso_8859-9" (characterSet) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso_8859-9" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso_8859-9" (charset) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso_8859-9:1989" (characterSet) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "iso_8859-9:1989" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "iso_8859-9:1989" (charset) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "l5" (characterSet) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "l5" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "l5" (charset) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "latin5" (characterSet) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "latin5" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "latin5" (charset) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "windows-1254" (characterSet) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "windows-1254" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "windows-1254" (charset) assert_equals: expected "WINDOWS-1254" but got "windows-1254"
FAIL Name "windows-1254" has label "x-cp1254" (characterSet) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "x-cp1254" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "x-cp1254" (charset) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1255" has label "cp1255" (characterSet) assert_equals: expected "WINDOWS-1255" but got "windows-1255"
FAIL Name "windows-1255" has label "cp1255" (inputEncoding) assert_equals: expected "WINDOWS-1255" but got "windows-1255"
FAIL Name "windows-1255" has label "cp1255" (charset) assert_equals: expected "WINDOWS-1255" but got "windows-1255"
FAIL Name "windows-1255" has label "windows-1255" (characterSet) assert_equals: expected "WINDOWS-1255" but got "windows-1255"
FAIL Name "windows-1255" has label "windows-1255" (inputEncoding) assert_equals: expected "WINDOWS-1255" but got "windows-1255"
FAIL Name "windows-1255" has label "windows-1255" (charset) assert_equals: expected "WINDOWS-1255" but got "windows-1255"
FAIL Name "windows-1255" has label "x-cp1255" (characterSet) assert_equals: expected "WINDOWS-1255" but got "ISO-8859-1"
FAIL Name "windows-1255" has label "x-cp1255" (inputEncoding) assert_equals: expected "WINDOWS-1255" but got "ISO-8859-1"
FAIL Name "windows-1255" has label "x-cp1255" (charset) assert_equals: expected "WINDOWS-1255" but got "ISO-8859-1"
FAIL Name "windows-1256" has label "cp1256" (characterSet) assert_equals: expected "WINDOWS-1256" but got "windows-1256"
FAIL Name "windows-1256" has label "cp1256" (inputEncoding) assert_equals: expected "WINDOWS-1256" but got "windows-1256"
FAIL Name "windows-1256" has label "cp1256" (charset) assert_equals: expected "WINDOWS-1256" but got "windows-1256"
FAIL Name "windows-1256" has label "windows-1256" (characterSet) assert_equals: expected "WINDOWS-1256" but got "windows-1256"
FAIL Name "windows-1256" has label "windows-1256" (inputEncoding) assert_equals: expected "WINDOWS-1256" but got "windows-1256"
FAIL Name "windows-1256" has label "windows-1256" (charset) assert_equals: expected "WINDOWS-1256" but got "windows-1256"
FAIL Name "windows-1256" has label "x-cp1256" (characterSet) assert_equals: expected "WINDOWS-1256" but got "ISO-8859-1"
FAIL Name "windows-1256" has label "x-cp1256" (inputEncoding) assert_equals: expected "WINDOWS-1256" but got "ISO-8859-1"
FAIL Name "windows-1256" has label "x-cp1256" (charset) assert_equals: expected "WINDOWS-1256" but got "ISO-8859-1"
FAIL Name "windows-1257" has label "cp1257" (characterSet) assert_equals: expected "WINDOWS-1257" but got "windows-1257"
FAIL Name "windows-1257" has label "cp1257" (inputEncoding) assert_equals: expected "WINDOWS-1257" but got "windows-1257"
FAIL Name "windows-1257" has label "cp1257" (charset) assert_equals: expected "WINDOWS-1257" but got "windows-1257"
FAIL Name "windows-1257" has label "windows-1257" (characterSet) assert_equals: expected "WINDOWS-1257" but got "windows-1257"
FAIL Name "windows-1257" has label "windows-1257" (inputEncoding) assert_equals: expected "WINDOWS-1257" but got "windows-1257"
FAIL Name "windows-1257" has label "windows-1257" (charset) assert_equals: expected "WINDOWS-1257" but got "windows-1257"
FAIL Name "windows-1257" has label "x-cp1257" (characterSet) assert_equals: expected "WINDOWS-1257" but got "ISO-8859-1"
FAIL Name "windows-1257" has label "x-cp1257" (inputEncoding) assert_equals: expected "WINDOWS-1257" but got "ISO-8859-1"
FAIL Name "windows-1257" has label "x-cp1257" (charset) assert_equals: expected "WINDOWS-1257" but got "ISO-8859-1"
FAIL Name "windows-1258" has label "cp1258" (characterSet) assert_equals: expected "WINDOWS-1258" but got "windows-1258"
FAIL Name "windows-1258" has label "cp1258" (inputEncoding) assert_equals: expected "WINDOWS-1258" but got "windows-1258"
FAIL Name "windows-1258" has label "cp1258" (charset) assert_equals: expected "WINDOWS-1258" but got "windows-1258"
FAIL Name "windows-1258" has label "windows-1258" (characterSet) assert_equals: expected "WINDOWS-1258" but got "windows-1258"
FAIL Name "windows-1258" has label "windows-1258" (inputEncoding) assert_equals: expected "WINDOWS-1258" but got "windows-1258"
FAIL Name "windows-1258" has label "windows-1258" (charset) assert_equals: expected "WINDOWS-1258" but got "windows-1258"
FAIL Name "windows-1258" has label "x-cp1258" (characterSet) assert_equals: expected "WINDOWS-1258" but got "ISO-8859-1"
FAIL Name "windows-1258" has label "x-cp1258" (inputEncoding) assert_equals: expected "WINDOWS-1258" but got "ISO-8859-1"
FAIL Name "windows-1258" has label "x-cp1258" (charset) assert_equals: expected "WINDOWS-1258" but got "ISO-8859-1"
FAIL Name "x-mac-cyrillic" has label "x-mac-cyrillic" (characterSet) assert_equals: expected "X-MAC-CYRILLIC" but got "x-mac-cyrillic"
FAIL Name "x-mac-cyrillic" has label "x-mac-cyrillic" (inputEncoding) assert_equals: expected "X-MAC-CYRILLIC" but got "x-mac-cyrillic"
FAIL Name "x-mac-cyrillic" has label "x-mac-cyrillic" (charset) assert_equals: expected "X-MAC-CYRILLIC" but got "x-mac-cyrillic"
FAIL Name "x-mac-cyrillic" has label "x-mac-ukrainian" (characterSet) assert_equals: expected "X-MAC-CYRILLIC" but got "x-mac-cyrillic"
FAIL Name "x-mac-cyrillic" has label "x-mac-ukrainian" (inputEncoding) assert_equals: expected "X-MAC-CYRILLIC" but got "x-mac-cyrillic"
FAIL Name "x-mac-cyrillic" has label "x-mac-ukrainian" (charset) assert_equals: expected "X-MAC-CYRILLIC" but got "x-mac-cyrillic"
PASS Name "gbk" has label "chinese" (characterSet) 
PASS Name "gbk" has label "chinese" (inputEncoding) 
PASS Name "gbk" has label "chinese" (charset) 
PASS Name "gbk" has label "csgb2312" (characterSet) 
PASS Name "gbk" has label "csgb2312" (inputEncoding) 
PASS Name "gbk" has label "csgb2312" (charset) 
PASS Name "gbk" has label "csiso58gb231280" (characterSet) 
PASS Name "gbk" has label "csiso58gb231280" (inputEncoding) 
PASS Name "gbk" has label "csiso58gb231280" (charset) 
PASS Name "gbk" has label "gb2312" (characterSet) 
PASS Name "gbk" has label "gb2312" (inputEncoding) 
PASS Name "gbk" has label "gb2312" (charset) 
FAIL Name "gbk" has label "gb_2312" (characterSet) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "gb_2312" (inputEncoding) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "gb_2312" (charset) assert_equals: expected "GBK" but got "ISO-8859-1"
PASS Name "gbk" has label "gb_2312-80" (characterSet) 
PASS Name "gbk" has label "gb_2312-80" (inputEncoding) 
PASS Name "gbk" has label "gb_2312-80" (charset) 
PASS Name "gbk" has label "gbk" (characterSet) 
PASS Name "gbk" has label "gbk" (inputEncoding) 
PASS Name "gbk" has label "gbk" (charset) 
PASS Name "gbk" has label "iso-ir-58" (characterSet) 
PASS Name "gbk" has label "iso-ir-58" (inputEncoding) 
PASS Name "gbk" has label "iso-ir-58" (charset) 
PASS Name "gbk" has label "x-gbk" (characterSet) 
PASS Name "gbk" has label "x-gbk" (inputEncoding) 
PASS Name "gbk" has label "x-gbk" (charset) 
PASS Name "gb18030" has label "gb18030" (characterSet) 
PASS Name "gb18030" has label "gb18030" (inputEncoding) 
PASS Name "gb18030" has label "gb18030" (charset) 
PASS Name "hz-gb-2312" has label "hz-gb-2312" (characterSet) 
PASS Name "hz-gb-2312" has label "hz-gb-2312" (inputEncoding) 
PASS Name "hz-gb-2312" has label "hz-gb-2312" (charset) 
PASS Name "big5" has label "big5" (characterSet) 
PASS Name "big5" has label "big5" (inputEncoding) 
PASS Name "big5" has label "big5" (charset) 
FAIL Name "big5" has label "big5-hkscs" (characterSet) assert_equals: expected "Big5" but got "Big5-HKSCS"
FAIL Name "big5" has label "big5-hkscs" (inputEncoding) assert_equals: expected "Big5" but got "Big5-HKSCS"
FAIL Name "big5" has label "big5-hkscs" (charset) assert_equals: expected "Big5" but got "Big5-HKSCS"
PASS Name "big5" has label "cn-big5" (characterSet) 
PASS Name "big5" has label "cn-big5" (inputEncoding) 
PASS Name "big5" has label "cn-big5" (charset) 
PASS Name "big5" has label "csbig5" (characterSet) 
PASS Name "big5" has label "csbig5" (inputEncoding) 
PASS Name "big5" has label "csbig5" (charset) 
PASS Name "big5" has label "x-x-big5" (characterSet) 
PASS Name "big5" has label "x-x-big5" (inputEncoding) 
PASS Name "big5" has label "x-x-big5" (charset) 
PASS Name "euc-jp" has label "cseucpkdfmtjapanese" (characterSet) 
PASS Name "euc-jp" has label "cseucpkdfmtjapanese" (inputEncoding) 
PASS Name "euc-jp" has label "cseucpkdfmtjapanese" (charset) 
PASS Name "euc-jp" has label "euc-jp" (characterSet) 
PASS Name "euc-jp" has label "euc-jp" (inputEncoding) 
PASS Name "euc-jp" has label "euc-jp" (charset) 
PASS Name "euc-jp" has label "x-euc-jp" (characterSet) 
PASS Name "euc-jp" has label "x-euc-jp" (inputEncoding) 
PASS Name "euc-jp" has label "x-euc-jp" (charset) 
PASS Name "iso-2022-jp" has label "csiso2022jp" (characterSet) 
PASS Name "iso-2022-jp" has label "csiso2022jp" (inputEncoding) 
PASS Name "iso-2022-jp" has label "csiso2022jp" (charset) 
PASS Name "iso-2022-jp" has label "iso-2022-jp" (characterSet) 
PASS Name "iso-2022-jp" has label "iso-2022-jp" (inputEncoding) 
PASS Name "iso-2022-jp" has label "iso-2022-jp" (charset) 
PASS Name "shift_jis" has label "csshiftjis" (characterSet) 
PASS Name "shift_jis" has label "csshiftjis" (inputEncoding) 
PASS Name "shift_jis" has label "csshiftjis" (charset) 
PASS Name "shift_jis" has label "ms932" (characterSet) 
PASS Name "shift_jis" has label "ms932" (inputEncoding) 
PASS Name "shift_jis" has label "ms932" (charset) 
PASS Name "shift_jis" has label "ms_kanji" (characterSet) 
PASS Name "shift_jis" has label "ms_kanji" (inputEncoding) 
PASS Name "shift_jis" has label "ms_kanji" (charset) 
PASS Name "shift_jis" has label "shift-jis" (characterSet) 
PASS Name "shift_jis" has label "shift-jis" (inputEncoding) 
PASS Name "shift_jis" has label "shift-jis" (charset) 
PASS Name "shift_jis" has label "shift_jis" (characterSet) 
PASS Name "shift_jis" has label "shift_jis" (inputEncoding) 
PASS Name "shift_jis" has label "shift_jis" (charset) 
PASS Name "shift_jis" has label "sjis" (characterSet) 
PASS Name "shift_jis" has label "sjis" (inputEncoding) 
PASS Name "shift_jis" has label "sjis" (charset) 
PASS Name "shift_jis" has label "windows-31j" (characterSet) 
PASS Name "shift_jis" has label "windows-31j" (inputEncoding) 
PASS Name "shift_jis" has label "windows-31j" (charset) 
PASS Name "shift_jis" has label "x-sjis" (characterSet) 
PASS Name "shift_jis" has label "x-sjis" (inputEncoding) 
PASS Name "shift_jis" has label "x-sjis" (charset) 
PASS Name "euc-kr" has label "cseuckr" (characterSet) 
PASS Name "euc-kr" has label "cseuckr" (inputEncoding) 
PASS Name "euc-kr" has label "cseuckr" (charset) 
PASS Name "euc-kr" has label "csksc56011987" (characterSet) 
PASS Name "euc-kr" has label "csksc56011987" (inputEncoding) 
PASS Name "euc-kr" has label "csksc56011987" (charset) 
PASS Name "euc-kr" has label "euc-kr" (characterSet) 
PASS Name "euc-kr" has label "euc-kr" (inputEncoding) 
PASS Name "euc-kr" has label "euc-kr" (charset) 
PASS Name "euc-kr" has label "iso-ir-149" (characterSet) 
PASS Name "euc-kr" has label "iso-ir-149" (inputEncoding) 
PASS Name "euc-kr" has label "iso-ir-149" (charset) 
PASS Name "euc-kr" has label "korean" (characterSet) 
PASS Name "euc-kr" has label "korean" (inputEncoding) 
PASS Name "euc-kr" has label "korean" (charset) 
PASS Name "euc-kr" has label "ks_c_5601-1987" (characterSet) 
PASS Name "euc-kr" has label "ks_c_5601-1987" (inputEncoding) 
PASS Name "euc-kr" has label "ks_c_5601-1987" (charset) 
PASS Name "euc-kr" has label "ks_c_5601-1989" (characterSet) 
PASS Name "euc-kr" has label "ks_c_5601-1989" (inputEncoding) 
PASS Name "euc-kr" has label "ks_c_5601-1989" (charset) 
PASS Name "euc-kr" has label "ksc5601" (characterSet) 
PASS Name "euc-kr" has label "ksc5601" (inputEncoding) 
PASS Name "euc-kr" has label "ksc5601" (charset) 
PASS Name "euc-kr" has label "ksc_5601" (characterSet) 
PASS Name "euc-kr" has label "ksc_5601" (inputEncoding) 
PASS Name "euc-kr" has label "ksc_5601" (charset) 
PASS Name "euc-kr" has label "windows-949" (characterSet) 
PASS Name "euc-kr" has label "windows-949" (inputEncoding) 
PASS Name "euc-kr" has label "windows-949" (charset) 
FAIL Name "replacement" has label "csiso2022kr" (characterSet) assert_equals: expected "REPLACEMENT" but got "ISO-2022-KR"
FAIL Name "replacement" has label "csiso2022kr" (inputEncoding) assert_equals: expected "REPLACEMENT" but got "ISO-2022-KR"
FAIL Name "replacement" has label "csiso2022kr" (charset) assert_equals: expected "REPLACEMENT" but got "ISO-2022-KR"
FAIL Name "replacement" has label "iso-2022-cn" (characterSet) assert_equals: expected "REPLACEMENT" but got "ISO-2022-CN"
FAIL Name "replacement" has label "iso-2022-cn" (inputEncoding) assert_equals: expected "REPLACEMENT" but got "ISO-2022-CN"
FAIL Name "replacement" has label "iso-2022-cn" (charset) assert_equals: expected "REPLACEMENT" but got "ISO-2022-CN"
FAIL Name "replacement" has label "iso-2022-cn-ext" (characterSet) assert_equals: expected "REPLACEMENT" but got "ISO-2022-CN-EXT"
FAIL Name "replacement" has label "iso-2022-cn-ext" (inputEncoding) assert_equals: expected "REPLACEMENT" but got "ISO-2022-CN-EXT"
FAIL Name "replacement" has label "iso-2022-cn-ext" (charset) assert_equals: expected "REPLACEMENT" but got "ISO-2022-CN-EXT"
FAIL Name "replacement" has label "iso-2022-kr" (characterSet) assert_equals: expected "REPLACEMENT" but got "ISO-2022-KR"
FAIL Name "replacement" has label "iso-2022-kr" (inputEncoding) assert_equals: expected "REPLACEMENT" but got "ISO-2022-KR"
FAIL Name "replacement" has label "iso-2022-kr" (charset) assert_equals: expected "REPLACEMENT" but got "ISO-2022-KR"
